Changed the for comprehension to use EitherT.fromEither.
